+++ src/client/components/905-icon.js
@@ -13,22 +13,20 @@ class Icon extends React.Component {
     icon:             React.PropTypes.string.isRequired,
     size:             React.PropTypes.string,   // lg, 2x, 3x, 4x, 5x
     fFixedWidth:      React.PropTypes.bool,
+    fSpin:            React.PropTypes.bool,
     style:            React.PropTypes.object,
     fDisabled:        React.PropTypes.bool,
     // all other props are passed through
   };
-  static defaultProps = {
-    fFixedWidth:      false,
-  };
 
-  render() { 
-    const { icon, size, fFixedWidth, fDisabled } = this.props;
+  render() {
+    const { icon, size, fFixedWidth, fSpin, fDisabled } = this.props;
     const otherProps = omit(this.props, ['icon', 'size', 'style']);
     if (fDisabled) otherProps.onClick = undefined;
     let className = `fa fa-${icon}`;
     if (size != null) className += ` fa-${size}`;
     if (fFixedWidth) className += ' fa-fw';
-    if (icon === 'circle-o-notch') className += " fa-spin"
+    if (icon === 'circle-o-notch' || fSpin) className += " fa-spin"
     return <i className={className} {...otherProps} style={style.icon(this.props)}/>;
   }
 }
@@ -37,6 +35,7 @@ const style = {
   icon: ({ fDisabled, style }) => merge({
     cursor: fDisabled ? undefined : 'pointer',
     color: fDisabled ? '#999' : undefined,
+    letterSpacing: 'normal',
   }, style),
 };