+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+* [m] Add spinner while parsing
+
 ## 0.1.5 (Apr. 16, 2016)
 
 * First public release